pyproject.toml
[project]
name = "sap-shell-checker-new-env"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nuitka>=2.8.6",
    "polars>=1.35.2",
    "python-docx>=1.2.0",
    "tqdm>=4.67.1",
    "xlsxwriter",
]
